refactor(storage): extract state trie overlay crate
