+++ .github/CODEOWNERS
@@ -38,6 +38,7 @@ crates/storage/libmdbx-rs/  @shekhirin
 crates/storage/nippy-jar/   @joshieDo @shekhirin
 crates/storage/provider/    @joshieDo @shekhirin
 crates/storage/storage-api/ @joshieDo
+crates/storage/storage-overlay/ @joshieDo @shekhirin
 crates/tasks/               @mattsse @DaniPopes
 crates/tokio-util/          @mattsse 
 crates/tracing/             @mattsse @shekhirin
