+++ Cargo.toml
@@ -106,6 +106,7 @@ members = [
     "crates/storage/nippy-jar/",
     "crates/storage/provider/",
     "crates/storage/storage-api/",
+    "crates/storage/storage-overlay/",
     "crates/tasks/",
     "crates/tokio-util/",
     "crates/tracing/",
@@ -413,6 +414,7 @@ reth-static-file = { path = "crates/static-file/static-file" }
 reth-static-file-types = { path = "crates/static-file/types", default-features = false }
 reth-storage-api = { path = "crates/storage/storage-api", default-features = false }
 reth-storage-errors = { path = "crates/storage/errors", default-features = false }
+reth-storage-overlay = { path = "crates/storage/storage-overlay" }
 reth-tasks = { path = "crates/tasks" }
 reth-testing-utils = { path = "testing/testing-utils" }
 reth-tokio-util = { path = "crates/tokio-util" }
