+++ crates/chain-state/Cargo.toml
@@ -18,9 +18,8 @@ reth-errors.workspace = true
 reth-execution-types.workspace = true
 reth-metrics.workspace = true
 reth-ethereum-primitives.workspace = true
-reth-primitives-traits = { workspace = true, features = ["dashmap"] }
+reth-primitives-traits.workspace = true
 reth-storage-api.workspace = true
-reth-tasks = { workspace = true, features = ["rayon"], optional = true }
 reth-trie-sparse = { workspace = true, features = ["std"] }
 reth-trie.workspace = true
 
@@ -42,7 +41,6 @@ derive_more.workspace = true
 metrics.workspace = true
 parking_lot.workspace = true
 pin-project.workspace = true
-rayon = { workspace = true, optional = true }
 serde = { workspace = true, optional = true }
 
 # optional deps for test-utils
@@ -81,7 +79,5 @@ test-utils = [
     "reth-primitives-traits/test-utils",
     "reth-trie/test-utils",
     "reth-ethereum-primitives/test-utils",
-    "reth-tasks?/test-utils",
     "reth-trie-sparse/test-utils",
 ]
-rayon = ["dep:rayon", "dep:reth-tasks"]
