+++ crates/engine/tree/src/launch.rs
@@ -13,7 +13,6 @@ use crate::{
     tree::{EngineApiTreeHandler, EngineValidator, TreeConfig, WaitForCaches},
 };
 use futures::Stream;
-use reth_chain_state::StateTrieOverlayManager;
 use reth_consensus::FullConsensus;
 use reth_engine_primitives::BeaconEngineMessage;
 use reth_evm::ConfigureEvm;
@@ -26,8 +25,8 @@ use reth_provider::{
 };
 use reth_prune::PrunerWithFactory;
 use reth_stages_api::{MetricEventsSender, Pipeline};
+use reth_storage_overlay::{ChangesetCache, StateTrieOverlayManager};
 use reth_tasks::Runtime;
-use reth_trie_db::ChangesetCache;
 use std::sync::Arc;
 
 /// Builds the engine [`ChainOrchestrator`] that drives the chain forward.
