+++ crates/storage/provider/Cargo.toml
@@ -20,6 +20,7 @@ reth-primitives-traits = { workspace = true, features = ["reth-codec", "secp256k
 reth-errors.workspace = true
 reth-storage-errors.workspace = true
 reth-storage-api = { workspace = true, features = ["std", "db-api"] }
+reth-storage-overlay.workspace = true
 reth-db = { workspace = true, features = ["mdbx"] }
 reth-db-api.workspace = true
 reth-prune-types.workspace = true
@@ -87,7 +88,10 @@ tokio-stream = { workspace = true, features = ["sync"] }
 
 [features]
 jemalloc = ["rocksdb/jemalloc"]
-partial-persistence = ["reth-stages-types/partial-persistence"]
+partial-persistence = [
+    "reth-stages-types/partial-persistence",
+    "reth-storage-overlay/partial-persistence",
+]
 test-utils = [
     "reth-db/test-utils",
     "reth-nippy-jar/test-utils",
