+++ crates/trie/db/Cargo.toml
@@ -20,16 +20,9 @@ reth-trie.workspace = true
 reth-trie-common.workspace = true
 reth-storage-api = { workspace = true, features = ["db-api"] }
 reth-storage-errors.workspace = true
-reth-metrics = { workspace = true, optional = true }
 
 # alloy
 alloy-primitives.workspace = true
-
-# misc
-parking_lot.workspace = true
-
-# metrics
-metrics = { workspace = true, optional = true }
 
 # tracing
 tracing.workspace = true
@@ -58,12 +51,11 @@ serde_json.workspace = true
 similar-asserts.workspace = true
 
 [features]
-metrics = ["reth-trie/metrics", "dep:reth-metrics", "dep:metrics"]
+metrics = ["reth-trie/metrics"]
 serde = [
     "similar-asserts/serde",
     "alloy-consensus/serde",
     "alloy-primitives/serde",
-    "parking_lot/serde",
     "reth-storage-api/serde",
     "reth-trie/serde",
     "reth-trie-common/serde",
