+++ crates/trie/parallel/Cargo.toml
@@ -50,7 +50,7 @@ reth-chainspec.workspace = true
 reth-ethereum-primitives.workspace = true
 reth-primitives-traits.workspace = true
 reth-provider = { workspace = true, features = ["test-utils"] }
-reth-trie-db.workspace = true
+reth-storage-overlay.workspace = true
 reth-trie = { workspace = true, features = ["test-utils"] }
 
 # misc
@@ -68,7 +68,6 @@ trie-debug = [
 test-utils = [
     "reth-primitives-traits/test-utils",
     "reth-provider/test-utils",
-    "reth-trie-db/test-utils",
     "reth-trie/test-utils",
     "reth-tasks/test-utils",
     "reth-chainspec/test-utils",
