+++ examples/custom-state-root/Cargo.toml
@@ -9,9 +9,9 @@ license.workspace = true
 reth-ethereum = { workspace = true, features = ["node", "test-utils"] }
 reth-engine-tree.workspace = true
 reth-evm.workspace = true
-reth-chain-state.workspace = true
 reth-primitives-traits.workspace = true
 reth-provider.workspace = true
+reth-storage-overlay.workspace = true
 reth-trie.workspace = true
 
 eyre.workspace = true
